# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vello Trade API
  description: |
    REST API for the Vello Trade B2B marketplace mobile app (Expo). JSON routes are mounted under **`/api`** (the mobile client uses `baseURL: {host}/api`).

    **Socket.IO** runs on the same HTTP server. After login, connect with JWT in handshake `auth: { token: "<jwt>" }`.
    Events include `room:join`, `chat:typing`, `message:new`, `chat:room:updated`. See the repository README for details.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, role]
              properties:
                email:
                  type: string
                  format: email
                phone:
                  type: string
                password:
                  type: string
                  minLength: 6
                role:
                  type: string
                  enum: [company, distributor, retail, end_user]
                name:
                  type: string
                businessName:
                  type: string
                address:
                  type: string
                companyId:
                  type: string
                distributorId:
                  type: string
      responses:
        "201":
          description: Created
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (email or phone + password; email may require OTP follow-up)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                email:
                  type: string
                phone:
                  type: string
                password:
                  type: string
                otpCode:
                  type: string
      responses:
        "200":
          description: Token or OTP challenge
  /api/auth/login-otp/request:
    post:
      tags: [Auth]
      summary: Request passwordless login OTP email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
  /api/auth/login-otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and receive JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                code:
                  type: string
  /api/auth/register/verify-email:
    post:
      tags: [Auth]
      summary: Verify end_user email after registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, otpCode]
              properties:
                email:
                  type: string
                otpCode:
                  type: string
  /api/auth/register/resend-email-code:
    post:
      tags: [Auth]
      summary: Resend registration verification code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
  /api/auth/forgot/request:
    post:
      tags: [Auth]
      summary: Request password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
  /api/auth/forgot/reset:
    post:
      tags: [Auth]
      summary: Reset password with code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, newPassword]
              properties:
                email:
                  type: string
                code:
                  type: string
                newPassword:
                  type: string
                  minLength: 6
  /api/auth/password:
    patch:
      tags: [Auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
...
  /api/auth/profile:
    get:
      tags: [Auth]
      summary: Current user (alias)
    patch:
      tags: [Auth]
      summary: Update display name only
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
  /api/auth/me/location:
    patch:
      tags: [Auth]
      summary: Set GeoJSON point on user for maps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                lng:
                  type: number
...